requirements.txt
django
djangorestframework
requests
pillow
